feat: new utilities files added

+++ scripts/lint.sh
@@ -1,30 +1,33 @@
 #!/bin/bash
 
 forPush="--for-push"
+isForPush="$( [[ $1 == $forPush ]]; echo $? )"
 
 cleanup() {
-  git stash pop -q
-  rm .script
-  echo "Stash Applied"
+  if [[ $isForPush == "0" ]]; then
+      git stash pop -q
+      rm .script
+      echo "Stash Applied"
+  fi
 }
 
 echo "Started"
 trap '[[ $? -ne 0 ]] && cleanup' EXIT
 
-if [[ $1 == $forPush ]]; then
+if [[ $isForPush == "0" ]]; then
     touch .script
     git stash push -uqm "Backup of auto commit functionality"
     echo "Stash Stored"
 fi
 
-yarn prettier:fix --log-level silent
+pnpm --silent prettier:fix --log-level silent
 echo "Prettier Completed"
-yarn lint:fix
+pnpm --silent lint:fix
 echo "Eslint Completed"
-yarn ts
+pnpm --silent ts
 echo "Typescript Completed"
 
-if [[ $1 == $forPush ]]; then
+if [[ $isForPush == "0" ]]; then
     git add .
     if ! git diff-index --quiet HEAD; then
       git commit -m "refactor: code reformatted" -q
